# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Indigo API Documentation
  description: "This is an API Documentation for the Indigo mibile and web application."
  contact:
    name: Nikethan
    url: https://www.linkedin.com/in/nikethan-selvanathan-a81928101
    email: nikesky273@gmail.com

# during dev, should point to your local machine
host: 127.0.0.1:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /branch/{ID}:
    x-swagger-router-controller: BranchDocumentController
    get:
      description: Get branch by ID
      summary: Get branch by ID
      tags:
        - Branches
      # used as the method name of the controller
      operationId: getBranchByID
      parameters:
        - name: ID
          in: path
          description: ID of branch
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BranchesDef"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /branch:
    x-swagger-router-controller: BranchDocumentController
    get:
      description: Retrieves all branches documents from the DB
      summary: Get all the branches documents of APIs
      tags:
        - Branches
      # used as the method name of the controller
      operationId: getAllBranches
      parameters:
        - name: limit
          in: query
          description: No of branches to return
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BranchesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a Branch document
      summary: Add a new Branch Document in DB
      tags: 
        - Branches
      # used as the method name of the controller
      operationId: createBranch
      parameters:
        - name: request
          in: body
          description: Request to the endpoint
          required: true
          schema:
            $ref: "#/definitions/BranchDef"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a Branch document
      summary: Update a Branch Document in DB
      tags: 
        - Branches
      # used as the method name of the controller
      operationId: updateBranch
      parameters:
        - name: branchID
          in: query
          description: Id of the branch Document
          required: true
          type: string
        - name: request
          in: body
          description: Request to the endpoint
          required: true
          schema:
            $ref: "#/definitions/BranchDef"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete a Branch document
      summary: Delete a Branch Document from DB
      tags: 
        - Branches
      # used as the method name of the controller
      operationId: deleteBranch
      parameters:
        - name: branchID
          in: query
          description: Id of the Branch Document
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  # Jobs API's
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    required:
      - id
    properties:
      id:
        type: string
...
  BranchesResponse:
    required:
      - branches
    properties:
      branches:
        type: array
        items:
          $ref: "#/definitions/BranchesDef"
...
  BranchesDef:
    properties:
      _id:
        type: string
      _rev:
        type: string
      organizationID:
        type: string
      branch:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      collection:
        type: string
      organization:
        items:
          $ref: "#/definitions/OrganizationsDef"
  OrganizationsDef:
    properties:
      _id:
        type: string
      _rev:
        type: string
      name:
        type: string
      image:
        type: string
      founder:
        type: string
      founded:
        type: string
      description:
        type: string
      collection:
        type: string
...
  BranchDef:
    properties:
      organizationID:
        type: string
      branch:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number